staching changes

+++ zebedee-cms/src/main/java/com/github/onsdigital/zebedee/api/Teams.java
@@ -4,14 +4,22 @@ import com.github.davidcarboni.restolino.framework.Api;
 import com.github.davidcarboni.restolino.helpers.Path;
 import com.github.onsdigital.zebedee.Zebedee;
 import com.github.onsdigital.zebedee.audit.Audit;
-import com.github.onsdigital.zebedee.exceptions.*;
-import com.github.onsdigital.zebedee.session.model.Session;
-import com.github.onsdigital.zebedee.teams.model.Team;
-import com.github.onsdigital.zebedee.teams.model.TeamList;
+import com.github.onsdigital.zebedee.exceptions.BadRequestException;
+import com.github.onsdigital.zebedee.exceptions.ConflictException;
+import com.github.onsdigital.zebedee.exceptions.ForbiddenException;
+import com.github.onsdigital.zebedee.exceptions.NotFoundException;
+import com.github.onsdigital.zebedee.exceptions.UnauthorizedException;
 import com.github.onsdigital.zebedee.model.Collection;
 import com.github.onsdigital.zebedee.model.KeyManager;
 import com.github.onsdigital.zebedee.service.ServiceSupplier;
+import com.github.onsdigital.zebedee.session.model.Session;
+import com.github.onsdigital.zebedee.session.service.Sessions;
+import com.github.onsdigital.zebedee.teams.model.Team;
+import com.github.onsdigital.zebedee.teams.model.TeamList;
+import com.github.onsdigital.zebedee.teams.service.TeamsService;
+import com.github.onsdigital.zebedee.user.model.User;
 import com.github.onsdigital.zebedee.user.service.UsersService;
+import org.apache.commons.lang3.StringUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -21,6 +29,9 @@ import javax.ws.rs.POST;
 import java.io.IOException;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
+
+import static com.github.onsdigital.zebedee.logging.CMSLogEvent.warn;
 
 /**
  * Created by thomasridd on 28/04/15.
@@ -36,12 +47,27 @@ import java.util.Set;
 @Api
 public class Teams {
 
+    private UsersService usersService;
+    private Sessions sessionsService;
+    private TeamsService teamsService;
+
     /**
      * Wrap static method calls to obtain service in function makes testing easier - class member can be
      * replaced with a mocked giving control of desired behaviour.
      */
     private ServiceSupplier<UsersService> usersServiceSupplier = () -> Root.zebedee.getUsersService();
 
+
+    public Teams() {
+        this.usersService = Root.zebedee.getUsersService();
+        this.sessionsService = Root.zebedee.getSessions();
+    }
+
+    Teams(final UsersService usersService, final Sessions sessionsService, TeamsService teamsService) {
+        this.usersService = usersService;
+        this.sessionsService = sessionsService;
+        this.teamsService = teamsService;
+    }
 
     /**
      * POST {@code /teams/[teamname]} creates a team with name {@code teamname}
@@ -147,17 +173,33 @@ public class Teams {
         return true;
     }
 
-    public boolean removeTeamMember(HttpServletRequest request, HttpServletResponse response) throws UnauthorizedException, IOException, NotFoundException, BadRequestException, ForbiddenException {
+    public boolean removeTeamMember(HttpServletRequest request, HttpServletResponse response)
+            throws UnauthorizedException, IOException, NotFoundException, BadRequestException, ForbiddenException {
 
         String teamName = getTeamName(request);
+        if (StringUtils.isEmpty(teamName)) {
+            return false;
+        }
+
+        Session session = getSession(request);
 
         Zebedee zebedee = Root.zebedee;
-        Session session = Root.zebedee.getSessions().get(request);
-        String email = request.getParameter("email");
-        Team team = zebedee.getTeamsService().findTeam(teamName);
+        Team team = getTeam(teamName);
+
+        String email = getEmailParam(request);
 
         zebedee.getTeamsService().removeTeamMember(email, team, session);
-        evaluateCollectionKeys(zebedee, session, team, email);
+
+        // If the user is a vewier remove key from them.
+        if (!zebedee.getPermissionsService().canEdit(email)) {
+            User user = Root.zebedee.getUsersService().getUserByEmail(email);
+            List<Collection> collections = getCollectionsAccessibleByTeam(team);
+
+            for (Collection col : collections) {
+                Root.zebedee.getCollectionKeyring().remove(user, col);
+            }
+
+        }
 
         Audit.Event.TEAM_MEMBER_REMOVED
                 .parameters()
@@ -167,6 +209,43 @@ public class Teams {
                 .actionedBy(session.getEmail())
                 .log();
         return true;
+    }
+
+    private Session getSession(HttpServletRequest request) throws UnauthorizedException {
+        Session session = null;
+        try {
+            session = sessionsService.get(request);
+        } catch (IOException ex) {
+            throw new UnauthorizedException(ex.getMessage());
+        }
+
+        if (session == null) {
+            throw new UnauthorizedException("unauthorised access session required but was null");
+        }
+        return session;
+    }
+
+    private Team getTeam(String name) throws IOException, NotFoundException {
+        Team team = null;
+        try {
+            team = teamsService.findTeam(name);
+        } catch (IOException | NotFoundException ex) {
+            throw ex;
+        }
+
+        if (team == null) {
+            warn().data("team", name).log("a team with the requested nane was not found");
+            throw new NotFoundException("requested team could not be found");
+        }
+        return team;
+    }
+
+    private String getEmailParam(HttpServletRequest request) throws BadRequestException {
+        String email = request.getParameter("email");
+        if (StringUtils.isEmpty(email)) {
+            throw new BadRequestException("email parameter required but none provided");
+        }
+        return email;
     }
 
     /**
@@ -180,9 +259,13 @@ public class Teams {
      * @throws NotFoundException
      * @throws BadRequestException
      */
-    private void evaluateCollectionKeys(Zebedee zebedee, Session session, Team team, String... emails) throws IOException, NotFoundException, BadRequestException, UnauthorizedException {
+    private void evaluateCollectionKeys(Zebedee zebedee, Session session, Team team, String... emails)
+            throws IOException, NotFoundException, BadRequestException, UnauthorizedException {
         for (Collection collection : zebedee.getCollections().list()) {
-            Set<Integer> teamIds = Root.zebedee.getPermissionsService().listViewerTeams(collection.description, session);
+            Set<Integer> teamIds = Root.zebedee
+                    .getPermissionsService()
+                    .listViewerTeams(collection.getDescription(), session);
+
             if (teamIds != null && teamIds.contains(team.getId())) {
                 for (String memberEmail : emails) {
                     KeyManager.distributeKeyToUser(zebedee, collection, session,
@@ -215,6 +298,13 @@ public class Teams {
         return result;
     }
 
+    private List<Collection> getCollectionsAccessibleByTeam(Team team) throws IOException {
+        return Root.zebedee.getCollections()
+                .list()
+                .stream()
+                .filter(c -> c.getDescription().getTeams().contains(team.getId()))
+                .collect(Collectors.toList());
+    }
 
     private static String getTeamName(HttpServletRequest request)
             throws IOException {
